requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
psycopg2-binary
python-dotenv
pandas
numpy
scipy
openai>=1.0.0
